♻️ refactor: 리뷰 반영 - 스트립트 코드 컴포넌트 분리
