+++ src/application/util/gtag.ts
@@ -1,9 +1,3 @@
-declare global {
-  interface Window {
-    gtag: any;
-  }
-}
-
 export const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GA_ID;
 
 // https://developers.google.com/analytics/devguides/collection/gtagjs/pages
