+++ src/infra/sdk/index.ts
@@ -1,2 +1,3 @@
 export * from "./channelIO";
+export * from "./google";
 export * from "./kakao";
